pyproject.toml
[project]
name = "v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "fastmcp>=2.8.1",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[mcp,openai]>=0.3.2",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
    "shap>=0.48.0",
    "uvicorn>=0.34.3",
]
